MCHMP-1 Introduce custom coroutine scope for task processing
# Motivation
GlobalScope - application level scope that might be full of its current workload and by this suspended for other clients - it increases latency. We separate this type of workload and make it independent by introducing new thread pool.

Being general purpose engine - it is hard to say which it is targeted - CPU or IO. So fixed thread pool should be fine.

# Other
- formatting;
- update .gitignore;
- fix tests - types compatibility;
- small refactoring.